+++ src/providers/BackgroundBlurProvider/index.tsx
@@ -88,10 +88,9 @@ const BackgroundBlurProvider: FC<Props> = ({ spec, options, children }) => {
   );
 
   useEffect(() => {
-    initializeBackgroundBlur();
     return () => {
       logger.info(
-        'Specs or options were changed. Destroying and re-initializing background blur processor.'
+        'Specs or options were changed. Destroying background blur processor.'
       );
       backgroundBlurProcessor?.destroy();
     };
@@ -148,6 +147,9 @@ const BackgroundBlurProvider: FC<Props> = ({ spec, options, children }) => {
         selectedDevice
       )}`
     );
+    // TODO: We don't need to intialize a new processor every time we create a background blur device
+    // We could potentially check for if a processor exists already AND that the processor isn't destroyed.
+    // If both of those statements are true, then chooseNewInnerDevice instead of creating a new processor
     const currentProcessor = await initializeBackgroundBlur();
     try {
       const logger =