+++ src/providers/VoiceFocusProvider/index.tsx
@@ -90,15 +90,9 @@ const VoiceFocusProvider: React.FC<Props> = ({
     logger.info(
       `Add Amazon Voice Focus to the following audio input device ${device}`
     );
-
-    if (voiceFocusDevice) {
-      const vf = await voiceFocusDevice.chooseNewInnerDevice(device);
-      logger.info(
-        'Re-used the same internal state to create an Amazon Voice Focus transform device.'
-      );
-      setVoiceFocusDevice(vf);
-      return vf;
-    }
+    // TODO: We don't need to intialize a new transformer every time we create a voice focus transformer device
+    // We could potentially check for if a transformer exists already AND that the voiceFocusDevice exists and hasnt been stopped.
+    // If both of those statements are true, then chooseNewInnerDevice instead of creating a new processor
 
     if (!isVoiceFocusSupported) {
       logger.debug('Not supported, not creating device.');
@@ -188,6 +182,10 @@ const VoiceFocusProvider: React.FC<Props> = ({
     createMeetingResponse: JoinMeetingInfo | undefined
   ) {
     // Throw away the old one and reinitialize.
+    voiceFocusDevice?.stop();
+    if (voiceFocusTransformer) {
+      VoiceFocusDeviceTransformer.destroyVoiceFocus(voiceFocusTransformer);
+    }
     setVoiceFocusTransformer(null);
     setVoiceFocusDevice(null);
     createVoiceFocusDeviceTransformer(
@@ -226,7 +224,31 @@ const VoiceFocusProvider: React.FC<Props> = ({
         `Current Amazon Voice Focus transform device: ${voiceFocusDevice}`
       );
     }
+    return () => {
+      if (voiceFocusDevice) {
+        logger.info(
+          'Destroying voice focus device : ' + JSON.stringify(voiceFocusDevice)
+        );
+        voiceFocusDevice?.stop();
+      } else {
+        logger.info("Voice focus device doesn't exist");
+      }
+    };
   }, [voiceFocusDevice]);
+
+  useEffect(() => {
+    return () => {
+      if (voiceFocusTransformer) {
+        VoiceFocusDeviceTransformer.destroyVoiceFocus(voiceFocusTransformer);
+        logger.info(
+          'Destroying voice focus transformer : ' +
+            JSON.stringify(voiceFocusTransformer)
+        );
+      } else {
+        logger.info("VoiceFocusTransformer doesn't exist");
+      }
+    };
+  }, [voiceFocusTransformer]);
 
   const value: VoiceFocusState = {
     isVoiceFocusSupported,