+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Changed
+- Updated `VoiceFocusProvider` to destroy the Voice Focus worker thread on unmount.
 
 ### Fixed